Add CLI for passing options.
Why: Need to be able to easily change the number of options.

How: Add a cli framework (fire) and use that to pass options.

Tags: cli, testing, fire
